checking the clustersize

+++ plugins/Infer2DCNN.cpp
@@ -469,10 +469,10 @@ public:
 				if(bigPixel) continue;
 				//printf("clustersize_x = %i, clustersize_y = %i\n",clustersize_x,clustersize_y);
 
-				if(clustersize_x%2==0) mid_x = int(clustersize_x/2)-1;
-				else mid_x = int(clustersize_x/2)-0.5;
-				if(clustersize_y%2==0) mid_y = int(clustersize_y/2)-1;
-				else mid_y = int(clustersize_y/2)-0.5;
+				if(clustersize_x%2==0) mid_x = clustersize_x/2-1;
+				else mid_x = clustersize_x/2-0.5;
+				if(clustersize_y%2==0) mid_y = clustersize_y/2-1;
+				else mid_y = clustersize_y/2-0.5;
 				int offset_x = 6 - mid_x;
 				int offset_y = 10 - mid_y;
 				//printf("offset_x = %i, offset_y = %i\n",offset_x,offset_y);
